pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "mdit-py-figure"
description = "markdown-it-py plugin to convert image paragraphs to HTML figure elements"
authors = [
    {name = "Mango Umbrella LLC", email = "hi@mangoumbrella.com"},
]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]
keywords = [
    "markdown",
    "markdown-it",
    "figure",
    "figcaption",
    "plugin",
]
requires-python = ">=3.10"
dependencies = [
    "markdown-it-py>=3.0.0",
]
dynamic = ["version"]

[project.urls]
Home = "https://github.com/mangoumbrella/mdit-py-figure"
Source = "https://github.com/mangoumbrella/mdit-py-figure"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/mdit_py_figure/_version.py"
template = '''
version = "{version}"
'''

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-benchmark>=5.1.0",
    "uv>=0.9.8",
]
